Fix bitwise operation between different enumeration types depr. warning
No idea why the inline operator defined in qt-cxx20-compat.h isn't used
here. It's used everywhere else.

+++ src/dialogs/expirydialog.cpp
@@ -128,7 +128,7 @@ private:
             auto okButton = buttonBox->button(QDialogButtonBox::Ok);
             KGuiItem::assign(okButton, KStandardGuiItem::ok());
             okButton->setDefault(true);
-            okButton->setShortcut(Qt::CTRL | Qt::Key_Return);
+            okButton->setShortcut(static_cast<int>(Qt::CTRL) | static_cast<int>(Qt::Key_Return));
             KGuiItem::assign(buttonBox->button(QDialogButtonBox::Cancel), KStandardGuiItem::cancel());
             qq->connect(buttonBox, &QDialogButtonBox::accepted, qq, &ExpiryDialog::accept);
             qq->connect(buttonBox, &QDialogButtonBox::rejected, qq, &QDialog::reject);
